Provide ability to build documentation
This commit creates a copy of the Makefile in the common/ directory
and places it in the doc-Service-Assurance-Framework directory, which
is expected to exist by the top level Makefile.

Updates built index name from foreman to saf. Adds a .gitignore to
avoid resulting build/ directory with generated contents from muddying
the repository. Updates top level Makefile to point at the appropriate
directory.