Provide auto building of documentation and publishing to gh-pages (#5)
* Provide documentation auto-build method

Provides a script and .travis.yml that allows for the automatic
building of documentation. When the branch is detected as being master
(merged down) then the script will also automatically push changes
into the gh-pages branch.

* Fail on WARN in asciidoc

Cause the documentation to fail on WARN which is the level that missing files
will show up in. We want the testing to fail if we're missing files that are
references in the master.adoc (for example).