Clean up how we're doing inline evaluation
Change the way inline evaluation is done to make things much cleaner. Adds the ifeval to the
stf-attributes.adoc file directly where parameters can be defined and used throughout the entire
book.

Moves to using ifdef and include_when_13 or include_when_16. Also centralizes the containerbin
parameter so it could be used in other procedures and assemblies.