Add OSP13 vs OSP16 content (#111)
* Add an html13 make target

Add an html13 Makefile target to make it easier to test builds for Queens

* Setup additional 13 vs 16 parametns

Line about enable-stf.yaml only built when OSP version is 16.0 or later. Builds < 16.0 use a different
value to containerbin than default (podman vs docker). All commands tested out of generated documentation.

* Drop procedure file no longer in use

* Clean up how we're doing inline evaluation

Change the way inline evaluation is done to make things much cleaner. Adds the ifeval to the
stf-attributes.adoc file directly where parameters can be defined and used throughout the entire
book.

Moves to using ifdef and include_when_13 or include_when_16. Also centralizes the containerbin
parameter so it could be used in other procedures and assemblies.

* Clean up code wrapping and quote hiding

* Multi-cloud configuration for OSP13

* Suggested changes from Martin

* We need EventPiplinePublishers to be empty

We need to make sure EventPiplinePulishers are set to an empty set so that
we don't create gnocchi and panko endpoints for sending events from Ceilometer