Clarify language around backends
The first sentence just sounds awkward to me, and in the second it's strange to use the plural of backends, and "include" when in each case it's a list of a singular item. I understand this is mirroring the layout of the api which is properly futureproof and allows for a dictionary of backends, but maybe we could could be more plain about it for now.