Updating qdstat --links output
The existing output was gathered from an OSP where collectd was configured to send messages in presettled mode. The use of this mode was eliminated, so users should expect to see a counter of 0 in the "presett(led)" column, and instead all of the messages should be counted in the "rel(eased)" column.