Fixed qdstat connections command
Existing command didn't work for me and seems invalid. The previous step was to look up the name of that pod specifically so it could be used in the `oc exec`. Not sure how the invalid command got it (or whether it's valid in a newer version of `oc`?) but this one works for sure.