Add clarification for backends persistent storage
Clarify that a dynamically created persistent volume is not required for
STF, but rather that persistent volumes are requested with a PVC, but
that statically defined volumes will still work. Add procedures and
links to show how to configure different storageClasses and the
requested volume size via PVC for the supported storage backends.