Add notes about duplicated sections
Add some hints in the documentation as a comment in case others come
along and make changes to these files. It's not possible due to style
guide rules to include partial procedure sections, so the contents needs
to be duplicated. In order to keep data from becoming out of sync the
notes are being added with points on where to make changes across both
files.