Update console output and add SGO sub step
Update the latest console output based on STF 1.3 installation
instructions.

Add explicit subscription step for Smart Gateway Operator to pull from
the stable-1.3 channel instead of relying on the defaultChannel in the
smart-gateway-operator package manifest, which after release of STF 1.4
is resulting in the deployment of the wrong CSV for Smart Gateway
Operator.

Resolves: rhbz#2046538