Provide documentation for all collectd plugins
Provide documentation for all supported or recommended collectd plugins
across the reference document. With these changes, all plugins are
either commented out and marked as Not Recommended or another note
indicating why documentation of that plugin does not apply currently.

In the future additional modules can be documented as requests come in
and are evaluated. For a complete set of collectd documentation please
see the official collectd documentation sources. This reference guide is
not intended to be an exhaustive set of documentation for collectd
plugins, only those that may be relevant to the operaion of an OpenStack
environment.

Resolves: rhbz#1923203