Provide documentation for all collectd plugins (#385)
* Provide documentation for all collectd plugins

Provide documentation for all supported or recommended collectd plugins
across the reference document. With these changes, all plugins are
either commented out and marked as Not Recommended or another note
indicating why documentation of that plugin does not apply currently.

In the future additional modules can be documented as requests come in
and are evaluated. For a complete set of collectd documentation please
see the official collectd documentation sources. This reference guide is
not intended to be an exhaustive set of documentation for collectd
plugins, only those that may be relevant to the operaion of an OpenStack
environment.

Resolves: rhbz#1923203

* Update common/collectd/ref_collectd-plugins.adoc

* Update common/collectd/ref_collectd-plugins.adoc

* Update common/collectd/ref_collectd-plugins.adoc

* Update common/collectd/ref_collectd-plugins.adoc

* Update common/collectd/ref_collectd-plugins.adoc

* Update common/collectd/ref_collectd-plugins.adoc

* Update common/collectd/ref_collectd-plugins.adoc

* Update common/collectd/ref_collectd-plugins.adoc

* Update common/collectd/ref_collectd-plugins.adoc

* Update common/collectd/ref_collectd-plugins.adoc

Co-authored-by: JoanneOFlynn2018 <45287002+JoanneOFlynn2018@users.noreply.github.com>

* Update common/collectd/ref_collectd-plugins.adoc

Co-authored-by: JoanneOFlynn2018 <45287002+JoanneOFlynn2018@users.noreply.github.com>

* Update common/collectd/ref_collectd-plugins.adoc

* Update common/collectd/ref_collectd-plugins.adoc

* Parameter is dynamically named

Co-authored-by: igallagh-redhat <62542106+igallagh-redhat@users.noreply.github.com>

* Update common/collectd/ref_collectd-plugins.adoc

Co-authored-by: igallagh-redhat <62542106+igallagh-redhat@users.noreply.github.com>

* Revert change to instances parameter in apache module

* Revert change to apache module header

* Fix syntax in tables

Fix syntax in tables and comment out amqp module which is not
recommended for use. (Also fixed name of module in case we make it
visible again in the future.)

* Fix apache module example

* Update common/collectd/ref_collectd-plugins.adoc

Co-authored-by: igallagh-redhat <62542106+igallagh-redhat@users.noreply.github.com>

* Document CollectdContainerAdditionalCapAdd for smart plugin

The smart plugin won't work if the system capability CAP_SYS_RAWIO isn't
enabled.

* Update common/collectd/ref_collectd-plugins.adoc

Co-authored-by: igallagh-redhat <62542106+igallagh-redhat@users.noreply.github.com>

* Provide Defaults column for default plugins

Create a |Defaults table column that contains the default values for all
plugins enabled by default via CollectdDefaultPlugins list. Default
values are those defined in the
<THT>/deployment/metrics/collectd-container-puppet.yaml file with
backfill from the puppet-collectd deployment tooling.

* Update common/collectd/ref_collectd-plugins.adoc

* Update common/collectd/ref_collectd-plugins.adoc

Co-authored-by: Leif Madsen <lmadsen@redhat.com>

* Call out default plugins more clearly

* Update reasons for plugins not being documented

* Add callout to disk explaining example

Add a callout to the disk plugin noting that the example provided is not
necessarily required, but is an example showing a filtering mechanism.
The default deployment requires no parameters.

* Address some issues from documentation review

* Update common/collectd/ref_collectd-plugins.adoc

Co-authored-by: Matthias Runge <mrunge@redhat.com>

Co-authored-by: JoanneOFlynn2018 <45287002+JoanneOFlynn2018@users.noreply.github.com>
Co-authored-by: igallagh-redhat <62542106+igallagh-redhat@users.noreply.github.com>
Co-authored-by: Matthias Runge <mrunge@redhat.com>