Fix up CRC deployment instructions
Clean up the CRC development environment deployment instructions.
Enabling the cluster monitoring within the environment is required due
to use of oauth-proxy, which is provided as part of the built in cluster
monitoring.

I also cleaned up the procedure a bit to not duplicate information and
remove procedure steps from the concept portion of the documentation.

This only affects the upstream documentation as this section is not
rendered downstream.