Remove use of OperatorHub.io for subscriptions
Remove the use of OperatorHub.io CatalogSource as a location for
Operator subscriptions. There is an updated version of the Promethueus
Operator available in the community-operators CatalogSource (which was
not the case previously), so I think migrating to using that helps
simplify installation instructions. There is an upgrade implication here
though, and we'll need to deal with that as well.