Add warning to not use procedure with gnocchi
Add a warning to not use the disable procedure when making use of the
Sending metrics to Gnocchi and Service Telemetry Framework procedure
since not all dependencies are provided as part of that instruction set,
since they are a super-set of the STF deployment instructions. In the
future we should probably just remove the Gnocchi deployment
instructions since we've re-written the autoscaling guide and that is
the one procedure that would provide Gnocchi deployments, and would
contain all the necessary dependencies in the provided THT environment
files.