Add procedure to disable services on OSP side (#407)
* Add procedure to disable services on OSP side

Add a procedure that disables the services provisioned when enabling STF.

Resolves: rhbz#2096853

* Add warning to not use procedure with gnocchi

Add a warning to not use the disable procedure when making use of the
Sending metrics to Gnocchi and Service Telemetry Framework procedure
since not all dependencies are provided as part of that instruction set,
since they are a super-set of the STF deployment instructions. In the
future we should probably just remove the Gnocchi deployment
instructions since we've re-written the autoscaling guide and that is
the one procedure that would provide Gnocchi deployments, and would
contain all the necessary dependencies in the provided THT environment
files.

* Apply suggestions from code review

Small changes from jof

* Update doc-Service-Telemetry-Framework/modules/proc_disabling-openstack-services-used-with-stf.adoc

Co-authored-by: JoanneOFlynn2018 <45287002+JoanneOFlynn2018@users.noreply.github.com>

* Update doc-Service-Telemetry-Framework/modules/proc_disabling-openstack-services-used-with-stf.adoc

Co-authored-by: JoanneOFlynn2018 <45287002+JoanneOFlynn2018@users.noreply.github.com>

* Update doc-Service-Telemetry-Framework/modules/proc_disabling-openstack-services-used-with-stf.adoc

Co-authored-by: JoanneOFlynn2018 <45287002+JoanneOFlynn2018@users.noreply.github.com>

Co-authored-by: JoanneOFlynn2018 <45287002+JoanneOFlynn2018@users.noreply.github.com>

(cherry picked from commit 20d6f11497c0b27053ac9a96369a225bb4187d37)