Build and push documentation for STF 1.5 (#409)
* Build and push documentation for STF 1.5

Update scripts and path references for STF 1.5 so that we can have
upstream documentation for this release. Going forward major changes
will happen in the main branch in preparation for STF 2.0. It will be
good to have a current version of documentation for the stable-1.5
branch which will reference the new release automation changes that
provides an upstream nightly build for stable-1.5 in its own index
image.

* Don't attempt to move non-existant file