Minor updates to dashboarding guide (#413)
* Minor updates to dashboarding guide

Perform some minor updates to the dashboarding guide, referencing
existing dashboards we have for virtual machine and memcached views.

Update the ServiceTelemetry manifest to reference the rhel8/grafana:7
container image which should provide more consistency in how things are
deployed, helping avoid a situation where newer versions of Grafana out
of hub.docker.com no longer interface with the version of Elasticsearch
that would be used when enabling events support by default.

Update the path to the dashboards being created to reference the
'stf-1/' directory to reduce confusion by no longer referencing stf-1.3
in the links.

Depends-On: https://github.com/infrawatch/dashboards/pull/50

* Update doc-Service-Telemetry-Framework/modules/con_dashboards.adoc

plural to single

* Update doc-Service-Telemetry-Framework/modules/proc_setting-up-grafana-to-host-the-dashboard.adoc

* Add note about STF 1.3 and revert path change

Revert path changes to the dashboards as use of symlinks as intended is
not possible. Instead, add a note stating that the reference to STF 1.3
is the earliest version compatible with the dashboards, and can be used
with STF versions 1.3 through 1.5.

Co-authored-by: JoanneOFlynn2018 <45287002+JoanneOFlynn2018@users.noreply.github.com>

(cherry picked from commit e23a1b360e3981645bebc8048f67d7bae33279ac)