Update link to STF life cycle page
Update the link to the STF life cycle page and remove reference to
supporting the two most recent versions of STF. This has recently
changed. STF 1.4 will be supported until the EOL of OpenShift 4.8 at
which point STF 1.4 will also be EOL. In the meantime STF 1.4 is now in
maintenance mode (CVE fixes only, no backport of features). STF 1.5 is
supported as of 4.10 and will EOL with RHOSP 17.1.