Add SNMP trap configuration parameters
Add SNMP trap configuration parameters that are being exposed via the
ServiceTelemetry object. Create a concept module that provides a better
overview of the OID configuration and link to location of the MIB
definition. Also provide an example prometheus rule that shows how to
override the OID value on a per-alert rule definition.

Closes: STF-1257