Clarify topic configuration in stf-connectors.yaml
Clarify the topic configuration in stf-connectors.yaml so the
administrator knows there is an underlying transposition happening where
the topic value is transposed into a topic address, and this is the full
address that would be referenced by the Smart Gateway configuration via
ServiceTelemetry object, clouds parameter.

Add a note that helps with understanding how the value is transposed
from the provided value into the full topic address listened on the bus.

Also clarify the difference between the Ceilometer configuration which
as a sub-parameter called topic, and the collectd configuration that has
a unique parameter name to identify the topic value.

Link back to an exaple configuration to make it clear how the value is
transposed into the full address. Update the linked clouds section to
match our unique-cloud configuration through the document.

Closes: rhbz#2189678
Closes: rhbz#2189677