Update OCP version support status
Update the version support status to specifically say that STF is
supported on OCP EUS releases. While the STF bundles are generated for a
range of releases, this is to support the ability of customers to
upgrade OCP clusters between EUS releases without needing to remove STF
first. Only minor testing is done against standard lifecycle releases of
OCP (odd-numbered minor releases).