Enable virt plugin's stats necessary for stf dashboard
This change overrides the default stats of virt plugin by enabling
just the ones that are needed for stf dashboard. This avoids
capturing the metrics that are not visualized in the dashboard.

Also, disable `vmem` plugin by default. There are no panels in stf
dashboard that visualize virtual memory subsystem of the kernel.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=2283987